Changes for the Order Booking API

+++ src/main/java/com/questionPro/groceryStore/repository/GroceryItemRepository.java
@@ -3,6 +3,6 @@ package com.questionPro.groceryStore.repository;
 import com.questionPro.groceryStore.entity.GroceryItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroceryItemRepository extends JpaRepository<GroceryItem, Long> {
+public interface GroceryItemRepository extends JpaRepository<GroceryItem, Integer> {
 
 }
